Support build tag to disable mocks and refactor code
