+++ capture/afpacket/afring/afring_mock_nodrain.go
@@ -1,3 +1,6 @@
+//go:build !slimcap_nomock
+// +build !slimcap_nomock
+
 package afring
 
 import (
