+++ event/event.go
@@ -1,3 +1,6 @@
+//go:build linux
+// +build linux
+
 package event
 
 import (
