+++ event/poll_default.go
@@ -10,6 +10,8 @@ import (
 )
 
 func pollBlock(fds *unix.PollFd, nfds int) unix.Errno {
+
+	// #nosec: G103
 	_, _, e := unix.Syscall6(unix.SYS_PPOLL, uintptr(unsafe.Pointer(fds)),
 		uintptr(nfds), uintptr(unsafe.Pointer(nil)), 0, 0, 0)
